+++ tests/model_multiple_pk.py
@@ -1,0 +1,10 @@
+"""
+This is the testing Models — Multiple PK
+"""
+from tortoise import fields
+from tortoise.models import Model
+
+
+class Tournament(Model):
+    id = fields.IntField(pk=True)
+    id2 = fields.IntField(pk=True)