+++ tortoise/__init__.py
@@ -277,8 +277,7 @@ class Tortoise:
             for model_name, model in app.items():
                 if model._meta._inited:
                     continue
-                else:
-                    model._meta._inited = True
+                model._meta._inited = True
                 if not model._meta.table:
                     model._meta.table = model.__name__.lower()
 
@@ -449,7 +448,7 @@ class Tortoise:
     def _get_config_from_config_file(cls, config_file: str) -> dict:
         _, extension = os.path.splitext(config_file)
         if extension in (".yml", ".yaml"):
-            import yaml
+            import yaml  # pylint: disable=C0415
 
             with open(config_file, "r") as f:
                 config = yaml.safe_load(f)