+++ tortoise/backends/base/client.py
@@ -39,6 +39,7 @@ class Capabilities:
         safe_indexes: bool = True,
         requires_limit: bool = False,
         inline_comment: bool = False,
+        supports_transactions: bool = True,
     ) -> None:
         super().__setattr__("_mutable", True)
 
@@ -47,6 +48,7 @@ class Capabilities:
         self.requires_limit = requires_limit
         self.safe_indexes = safe_indexes
         self.inline_comment = inline_comment
+        self.supports_transactions = supports_transactions
 
         super().__setattr__("_mutable", False)