+++ tortoise/backends/base/schema_generator.py
@@ -1,5 +1,4 @@
 import logging
-import warnings
 from typing import List, Optional, Set
 
 from tortoise import fields
@@ -46,9 +45,9 @@ class BaseSchemaGenerator:
     def _create_string(
         self, db_field: str, field_type: str, nullable: str, unique: str, is_pk: bool, comment: str
     ) -> str:
-        # children can override this function to customize thier sql queries
+        # children can override this function to customize their sql queries
 
-        field_creation_string = self.FIELD_TEMPLATE.format(
+        return self.FIELD_TEMPLATE.format(
             name=db_field,
             type=field_type,
             nullable=nullable,
@@ -57,7 +56,12 @@ class BaseSchemaGenerator:
             primary=" PRIMARY KEY" if is_pk else "",
         ).strip()
 
-        return field_creation_string
+    def _create_fk_string(
+        self, db_field: str, table: str, field: str, on_delete: str, comment: str
+    ) -> str:
+        return self.FK_TEMPLATE.format(
+            db_field=db_field, table=table, field=field, on_delete=on_delete, comment=comment
+        )
 
     def _get_primary_key_create_string(
         self, field_object: fields.Field, field_name: str, comment: str
@@ -91,6 +95,9 @@ class BaseSchemaGenerator:
     def _table_generate_extra(self, table: str) -> str:  # pylint: disable=R0201
         return ""
 
+    def _get_inner_statements(self) -> List[str]:
+        return []
+
     @staticmethod
     def _make_hash(*args: str, length: int) -> str:
         # Hash a set of string values and get a digest of the given length.
@@ -122,10 +129,13 @@ class BaseSchemaGenerator:
 
     def _get_field_type(self, field_object) -> str:
         field_object_type = type(field_object)
-        while field_object_type.__bases__ and field_object_type not in self.FIELD_TYPE_MAP:
+        while (
+            field_object_type.__bases__
+            and field_object_type not in self.FIELD_TYPE_MAP  # type: ignore
+        ):
             field_object_type = field_object_type.__bases__[0]
 
-        field_type = self.FIELD_TYPE_MAP[field_object_type]
+        field_type = self.FIELD_TYPE_MAP[field_object_type]  # type: ignore
 
         if isinstance(field_object, fields.DecimalField):
             field_type = field_type.format(field_object.max_digits, field_object.decimal_places)
@@ -175,7 +185,8 @@ class BaseSchemaGenerator:
                     unique=unique,
                     is_pk=field_object.pk,
                     comment="",
-                ) + self.FK_TEMPLATE.format(
+                ) + self._create_fk_string(
+                    db_field=db_field,
                     table=field_object.reference.type._meta.table,
                     field=field_object.reference.type._meta.db_pk_field,
                     on_delete=field_object.reference.on_delete,
@@ -217,6 +228,18 @@ class BaseSchemaGenerator:
 
             fields_to_create.extend(unique_together_sqls)
 
+        # Indexes.
+        field_indexes_sqls = [
+            val
+            for val in [
+                self._get_index_sql(model, [field_name], safe=safe)
+                for field_name in fields_with_index
+            ]
+            if val
+        ]
+
+        fields_to_create.extend(self._get_inner_statements())
+
         table_fields_string = "\n    {}\n".format(",\n    ".join(fields_to_create))
         table_comment = (
             self._table_comment_generator(
@@ -234,20 +257,7 @@ class BaseSchemaGenerator:
             extra=self._table_generate_extra(table=model._meta.table),
         )
 
-        # Indexes.
-        field_indexes_sqls = [
-            self._get_index_sql(model, [field_name], safe=safe) for field_name in fields_with_index
-        ]
-        if safe and not self.client.capabilities.safe_indexes:
-            warnings.warn(
-                f"Skipping creation of field indexes: safe index creation is not supported"
-                f" yet for {self.client.capabilities.dialect}. Please find the SQL queries"
-                f" to create the indexes in the logs."
-            )
-            for fis in field_indexes_sqls:
-                logger.warning(fis)
-        else:
-            table_create_string = "\n".join([table_create_string, *field_indexes_sqls])
+        table_create_string = "\n".join([table_create_string, *field_indexes_sqls])
 
         table_create_string += self._post_table_hook()