+++ tortoise/fields.py
@@ -3,7 +3,7 @@ import functools
 import json
 import uuid
 from decimal import Decimal
-from typing import TYPE_CHECKING, Any, Optional, Union
+from typing import TYPE_CHECKING, Any, Optional, Type, Union
 from uuid import UUID
 
 import ciso8601
@@ -13,6 +13,7 @@ from tortoise.exceptions import ConfigurationError, NoValuesFetched, Operational
 
 if TYPE_CHECKING:  # pragma: nocoverage
     from tortoise.models import Model
+
 
 CASCADE = "CASCADE"
 RESTRICT = "RESTRICT"
@@ -24,13 +25,26 @@ JSON_DUMPS = functools.partial(json.dumps, separators=(",", ":"))
 JSON_LOADS = json.loads
 
 
-class Field:
+class _FieldMeta(type):
+    def __new__(mcs, name, bases, attrs):
+        if len(bases) > 1:
+            # Instantiate class with only the 1st base class (should be Field)
+            cls = type.__new__(mcs, name, (bases[0],), attrs)  # type: Type[Field]
+            # All other base classes are our meta types, we store them in class attributes
+            cls.type = bases[1] if len(bases) == 2 else bases[1:]
+            return cls
+        return type.__new__(mcs, name, bases, attrs)
+
+
+class Field(metaclass=_FieldMeta):
     """
     Base Field type.
     """
 
+    # Type is a readonly property for the instance, it is set by _FieldMeta
+    type = None  # type: Type[Any]
+
     __slots__ = (
-        "type",
         "source_field",
         "generated",
         "pk",
@@ -45,9 +59,12 @@ class Field:
     )
     has_db_field = True
 
+    # This method is just to make IDE happy
+    def __new__(cls, *args, **kwargs):
+        return super().__new__(cls)
+
     def __init__(
         self,
-        type=None,  # pylint: disable=W0622
         source_field: Optional[str] = None,
         generated: bool = False,
         pk: bool = False,
@@ -60,7 +77,6 @@ class Field:
         description: Optional[str] = None,
         **kwargs,
     ) -> None:
-        self.type = type
         self.source_field = source_field
         self.generated = generated
         self.pk = pk
@@ -76,19 +92,19 @@ class Field:
     def to_db_value(self, value: Any, instance) -> Any:
         if value is None or type(value) == self.type:  # pylint: disable=C0123
             return value
-        return self.type(value)
+        return self.type(value)  # pylint: disable=E1102
 
     def to_python_value(self, value: Any) -> Any:
         if value is None or isinstance(value, self.type):
             return value
-        return self.type(value)
+        return self.type(value)  # pylint: disable=E1102
 
     @property
     def required(self):
         return self.default is None and not self.null and not self.generated
 
 
-class IntField(Field):
+class IntField(Field, int):
     """
     Integer field. (32-bit signed)
 
@@ -99,10 +115,10 @@ class IntField(Field):
     def __init__(self, pk: bool = False, **kwargs) -> None:
         if pk:
             kwargs["generated"] = bool(kwargs.get("generated", True))
-        super().__init__(int, pk=pk, **kwargs)
+        super().__init__(pk=pk, **kwargs)
 
 
-class BigIntField(Field):
+class BigIntField(Field, int):
     """
     Big integer field. (64-bit signed)
 
@@ -113,10 +129,10 @@ class BigIntField(Field):
     def __init__(self, pk: bool = False, **kwargs) -> None:
         if pk:
             kwargs["generated"] = bool(kwargs.get("generated", True))
-        super().__init__(int, pk=pk, **kwargs)
+        super().__init__(pk=pk, **kwargs)
 
 
-class SmallIntField(Field):
+class SmallIntField(Field, int):
     """
     Small integer field. (16-bit signed)
 
@@ -129,10 +145,10 @@ class SmallIntField(Field):
     def __init__(self, pk: bool = False, **kwargs) -> None:
         if pk:
             kwargs["generated"] = bool(kwargs.get("generated", True))
-        super().__init__(int, pk=pk, **kwargs)
+        super().__init__(pk=pk, **kwargs)
 
 
-class CharField(Field):
+class CharField(Field, str):  # type: ignore[misc]  # noqa
     """
     Character field.
 
@@ -148,18 +164,15 @@ class CharField(Field):
         if int(max_length) < 1:
             raise ConfigurationError("'max_length' must be >= 1")
         self.max_length = int(max_length)
-        super().__init__(str, **kwargs)
+        super().__init__(**kwargs)
 
 
-class TextField(Field):
+class TextField(Field, str):  # type: ignore[misc]  # noqa
     """
     Large Text field.
     """
 
     __slots__ = ()
-
-    def __init__(self, **kwargs) -> None:
-        super().__init__(str, **kwargs)
 
 
 class BooleanField(Field):
@@ -169,11 +182,11 @@ class BooleanField(Field):
 
     __slots__ = ()
 
-    def __init__(self, **kwargs) -> None:
-        super().__init__(bool, **kwargs)
+    # Bool is not subclassable, so we specify type here
+    type = bool
 
 
-class DecimalField(Field):
+class DecimalField(Field, Decimal):
     """
     Accurate decimal field.
 
@@ -192,12 +205,12 @@ class DecimalField(Field):
             raise ConfigurationError("'max_digits' must be >= 1")
         if int(decimal_places) < 0:
             raise ConfigurationError("'decimal_places' must be >= 0")
-        super().__init__(Decimal, **kwargs)
+        super().__init__(**kwargs)
         self.max_digits = max_digits
         self.decimal_places = decimal_places
 
 
-class DatetimeField(Field):
+class DatetimeField(Field, datetime.datetime):
     """
     Datetime field.
 
@@ -215,7 +228,7 @@ class DatetimeField(Field):
     def __init__(self, auto_now: bool = False, auto_now_add: bool = False, **kwargs) -> None:
         if auto_now_add and auto_now:
             raise ConfigurationError("You can choose only 'auto_now' or 'auto_now_add'")
-        super().__init__(datetime.datetime, **kwargs)
+        super().__init__(**kwargs)
         self.auto_now = auto_now
         self.auto_now_add = auto_now | auto_now_add
 
@@ -238,15 +251,12 @@ class DatetimeField(Field):
         return value
 
 
-class DateField(Field):
+class DateField(Field, datetime.date):
     """
     Date field.
     """
 
     __slots__ = ()
-
-    def __init__(self, **kwargs) -> None:
-        super().__init__(datetime.date, **kwargs)
 
     def to_python_value(self, value: Any) -> Optional[datetime.date]:
         if value is None or isinstance(value, datetime.date):
@@ -254,15 +264,12 @@ class DateField(Field):
         return ciso8601.parse_datetime(value).date()
 
 
-class TimeDeltaField(Field):
+class TimeDeltaField(Field, datetime.timedelta):
     """
     A field for storing time differences.
     """
 
     __slots__ = ()
-
-    def __init__(self, **kwargs) -> None:
-        super().__init__(datetime.timedelta, **kwargs)
 
     def to_python_value(self, value: Any) -> Optional[datetime.timedelta]:
         if value is None or isinstance(value, datetime.timedelta):
@@ -275,7 +282,7 @@ class TimeDeltaField(Field):
         return (value.days * 86400000000) + (value.seconds * 1000000) + value.microseconds
 
 
-class FloatField(Field):
+class FloatField(Field, float):
     """
     Float (double) field.
     """
@@ -283,10 +290,10 @@ class FloatField(Field):
     __slots__ = ()
 
     def __init__(self, **kwargs) -> None:
-        super().__init__(float, **kwargs)
+        super().__init__(**kwargs)
 
 
-class JSONField(Field):
+class JSONField(Field, dict, list):  # type: ignore[misc]  # noqa
     """
     JSON field.
 
@@ -301,7 +308,7 @@ class JSONField(Field):
     __slots__ = ("encoder", "decoder")
 
     def __init__(self, encoder=JSON_DUMPS, decoder=JSON_LOADS, **kwargs) -> None:
-        super().__init__(type=(dict, list), **kwargs)
+        super().__init__(**kwargs)
         self.encoder = encoder
         self.decoder = decoder
 
@@ -313,12 +320,12 @@ class JSONField(Field):
     def to_python_value(
         self, value: Optional[Union[str, dict, list]]
     ) -> Optional[Union[dict, list]]:
-        if value is None or isinstance(value, self.type):
+        if value is None or isinstance(value, (dict, list)):
             return value
         return self.decoder(value)
 
 
-class UUIDField(Field):
+class UUIDField(Field, UUID):
     """
     UUID Field
 
@@ -331,7 +338,7 @@ class UUIDField(Field):
         if kwargs.get("pk", False):
             if "default" not in kwargs:
                 kwargs["default"] = uuid.uuid4
-        super().__init__(type=UUID, **kwargs)
+        super().__init__(**kwargs)
 
     def to_db_value(self, value: Any, instance) -> Optional[str]:
         if value is None:
@@ -374,7 +381,12 @@ class ForeignKeyField(Field):
                 Can only be set is field has a ``default`` set.
     """
 
-    __slots__ = ("model_name", "related_name", "on_delete")
+    __slots__ = (
+        "type",  # type will be set later, so we need a slot to be able to write it
+        "model_name",
+        "related_name",
+        "on_delete",
+    )
     has_db_field = False
 
     def __init__(
@@ -419,6 +431,7 @@ class ManyToManyField(Field):
     """
 
     __slots__ = (
+        "type",  # Here we need type to be able to set dyamically
         "model_name",
         "related_name",
         "forward_key",
@@ -435,9 +448,11 @@ class ManyToManyField(Field):
         forward_key: Optional[str] = None,
         backward_key: str = "",
         related_name: str = "",
+        type=None,
         **kwargs,
     ) -> None:
         super().__init__(**kwargs)
+        self.type = type
         if len(model_name.split(".")) != 2:
             raise ConfigurationError('Foreign key accepts model name in format "app.Model"')
         self.model_name = model_name
@@ -449,7 +464,7 @@ class ManyToManyField(Field):
 
 
 class BackwardFKRelation(Field):
-    __slots__ = ("type", "relation_field")
+    __slots__ = ("type", "relation_field")  # Here we need type to be able to set dyamically
     has_db_field = False
 
     def __init__(
@@ -459,7 +474,8 @@ class BackwardFKRelation(Field):
         null: bool,
         description: Optional[str],
     ) -> None:
-        super().__init__(type=type, null=null)
+        super().__init__(null=null)
+        self.type = type
         self.relation_field = relation_field
         self.description = description